Move TreeReader to the correct namespace
