+++ SAGESharp.Tests/SAGESharp.Tests.csproj
@@ -87,6 +87,7 @@
     <Compile Include="IO\Binary\ReaderTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBinarySerializerTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeBuilderTests.cs" />
+    <Compile Include="IO\Binary\TreeBasedSerialization\TreeReaderTests.cs" />
     <Compile Include="IO\Binary\TreeBasedSerialization\TreeWriterTests.cs" />
     <Compile Include="IO\Binary\WriterTests.cs" />
     <Compile Include="IO\EdgeTests.cs" />
@@ -96,7 +97,6 @@
     <Compile Include="IO\PrimitiveTypeDataNodeTests.cs" />
     <Compile Include="IO\StringDataNodeTests.cs" />
     <Compile Include="IO\Substitutes.cs" />
-    <Compile Include="IO\TreeReaderTests.cs" />
     <Compile Include="IO\Trees\Builders.cs" />
     <Compile Include="IO\Trees\TreeWithHeight1.cs" />
     <Compile Include="IO\Trees\TreeWithHeight2.cs" />
