+++ SAGESharp/IO/Binary/BinarySerializer.cs
@@ -4,6 +4,7 @@
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
 using NUtils.Validations;
+using SAGESharp.IO.Binary.TreeBasedSerialization;
 using System;
 using System.Collections.Generic;
 using System.IO;
@@ -26,10 +27,10 @@ namespace SAGESharp.IO.Binary
         /// <returns>A binary serializer for type <typeparamref name="T"/>.</returns>
         public static IBinarySerializer<T> ForType<T>()
         {
-            return new TreeBasedSerialization.TreeBinarySerializer<T>(
+            return new TreeBinarySerializer<T>(
                 treeReader: treeReader.Read,
                 treeWriter: WriteTree,
-                rootNode: TreeBasedSerialization.TreeBuilder.BuildTreeForType(typeof(T)),
+                rootNode: TreeBuilder.BuildTreeForType(typeof(T)),
                 footerAligner: AlignFooter
             );
         }
